skills/espionar-concorrente-pro/scripts/requirements.txt
playwright>=1.40
google-genai>=0.3.0
jinja2>=3.1
httpx>=0.27
python-dotenv>=1.0
